cleanup: remove unused AccountHelper#isAccount predicate.
The logic here wasn't completely correct. A BLS account in FEVM
would look like an account. It'll return true for precompiles.
An inexistent address (which could later become an account) will
return false.

Not sure if these behaviours were intended, since the method is
not documented. But given that it's also not used, it's better to
drop it so it doesn't get used in this form.